feat: implement the `Name` object (#26)
This PR adds implementation of the `Name` object. The goal of this
object is to make it easier to encode keys of a `Dictionary` object, but
also any other occurrence of the `Name`.

The intention was to implement a `Dictionary` object which can be used
internally in other objects. However, such design would not allow for
good type safety in implementation, and we decided to implement helper
types that make creation of `Dictionary` objects easier instead.

Closes #24