feat: implement `Catalog` and PDF generation flow (#31)
In this PR we implement the `Catalog` PDF object. This is the piece that
ties the whole document together. Basically, Catalog defines where the
root page tree node is, and root page tree node holds the pages (and
eventually other page trees).

Because of this, the PDF generation flow is somewhat reworked in this PR
as well. The `ObjRef` is renamed to `ObjId` as that carries the meaning
better. There is now a single `IdManager` class that handles the
generation of IDs. At the moment it's possible to create `IdManager`
from anywhere using its `Default` implementation. We should think about
how to restrict this, as it's probably better idea to let the `Document`
handle IDs (to have a single source of truth for the IDs).

Closes #6
Closes #12