refactor: improve current `pdfgen` API (#43)
This PR aims to improve the current API of `pdfgen`. In particular,
`Document` can now be built by using a builder. This adds easy support
to modify defaults in a `Document`.

Taking leverage of this, it's now possible to set default page size at
root `PageTree` node which is inherited by other pages. This means that
we don't have to specify the page size every time we want to create a
page.

## Integration testing

I wanted to test the new API, and one of the possible ways to do this is
to take a look at the generated PDF file. However, `insta` does not have
great UX when generating arbitrary snapshot files. It does support
binary format, but no diffing for that.
So I decided to try `pretty_assertions` for this, where we have a lot
more control on how and what files we store etc. So this PR adds basic
setup for integration testing as well.