feat: implement basic support for the `Stream` object (#49)
### What?
This adds support for `Stream` PDF object.

### Why?
This is required for many subsequent tasks, as any type of content most
likely is to be encoded as `Content Stream` object, which is a special
case of `Stream` object.

### How?
Implementation is very simple, uses `Vec<u8>` internally to handle the
bytes. Alongside of that we support only the `Length` field in the
`Stream`'s dictionary.

### Notes
Note that `Content Stream` object is somewhat abstract. Each content
type has it's own specific dictionary entries, but the bytes rendering
is same for all of them.

### Checklist
- [x] Tests are added/updated.
- [x] Documentation is added/updated.
- [x] Self-review of code changes completed.

### Related Issues
Closes #35