feat: implement support for basic UTF8 encoded PDF String (#50)
### What?
Add support for very basic UTF8 encoded PDF String. 

### Why?
This is one of the requirements for #39. 

### How?
For now this is mostly a simple wrapper that wraps Rust's own
[`String`](https://doc.rust-lang.org/std/string/struct.String.html)
type. It can be created once, cannot be cloned and can be written to the
PDF.

### Notes
PDF Supports more encodings, such as ASCII String, UTF16-BE encoding
etc. We will focus on UTF8 encoding **only** at least for now. This is
very basic support for now, all text manipulation should be done on the
Rust's string type before finally converting it into `PdfString` and
writing it into a document.

### Checklist
- [x] Tests are added/updated.
- [x] Documentation is added/updated.
- [x] Self-review of code changes completed.

### Related Issues
Closes #36