feat: implement support for raster images (#51)
### What?
This PR adds support for raster image content in PDF. This is very basic
support, but should be enough to display a simple image.

### Why?
It's one of the required features, see #41. 

### How?
There are three parts for images in PDF: 

1. Add the image dictionary and image stream to the PDF
2. Put the image into Page's resources
3. Add image drawing operators to the Page's content stream

It is not clear what the best way to handle this is. For now, image
loading is done through the `Document` struct. This allows us to add the
Image object to the Document, handling the first part. Loaded image can
be adjusted, and then added to the page. Page takes only a `ObjId` of
the image, which is not super type safe but allows more flexibility in
the API. Page updates it's resource dictionary (handling the second
part), and finally Page must handle rendering of it's content stream.
Rendering of the content stream is the final part, but it's still not
clear what the best way to do this is - probably part of encoding the
page itself?

### Notes
This PR has a lot of refactoring. For example, `Object` trait takes now
`&mut dyn Write` which is a little slower than `&mut impl Write`, but
makes it
[`dyn-compatible`](https://doc.rust-lang.org/book/ch17-02-trait-objects.html).

### Checklist
- [x] Tests are added/updated.
- [x] Documentation is added/updated.
- [x] Self-review of code changes completed.

### Related Issues
Closes #41