refactor: remove `objs` field from `Document`
The objects will be handled in page as they are closely coupled to page.
Objects are part of resources for a page, and hence the page should
handle ownership of such objects. Some specific objects can be shared,
such as fonts, and they should be handled globally in the `Document`.