refactor: improve API of `pdfgen` (#55)
### What?
This PR adds multiple changes that should enable more ergonomic usage of
the `pdfgen` crate. In particular, we want to make it easy to add
objects to a page, handle object IDs and referencing easier etc.

### Why?
It was previously somewhat hard to hold a reference to a page and load
an image for example. This was because all operations were tied to the
document. However, that is not always necessary. Users should be able to
get a reference to a page, then load an image and add it to the page,
then load a font and add it to the page etc. With the new approach,
loading of image etc. is a fully independent operation.

This should de-couple the ID of objects from a particular object. We
should be able to add the same object to multiple documents without
needing to re-create them.

### How?
Objects such as Images do not store an ID, otherwise they are completely
tied to a particular Document with particular order of operations. But
they still need to be referenced somehow, and this is now handled at
appropriate positions. Right now, only `Page` needs such handling, and
this is done by creating a list of `Renderable`s from `Resources` which
allocates IDs for resources. These `Renderable`s can then be used both
for generating references to an entry and for writing the entry itself
(e.g. `Image` dictionary and its stream).

### Notes
There's a lot of restructuring in this PR. In particular, `ObjId`,
`IdManager`, `Document` and `DocumentBuilder` are moved in such way that
we can expose `IdManager::clone()` __only__ to `Document`. This is
needed, because the `Document::write` method now creates `ObjId`s as
needed. If we didn't clone `IdManager` inside this function, we would
generate a different document on multiple consecutive writes. One nice
side-effect of this change is that now a document can be written in
parallel to different files for example.

Also, this ID handling is for now implemented only in `Page`. IDs for
objects stored in `Document` directly is still not solved in a great
way. One particular option would be to have a wrapper for object that is
stored in Document and then use this. For example:

```rust
struct GlobalObject<T> {
    id: ObjId, // Created by `Document`
    obj: T,
}

// then in page for example: 
pub struct Page {
    // ... other fields
    global_fonts: Vec<GlobalObject<Font>>,
}
```

Such pattern would enable us to distinguish page-local objects and
objects that are stored directly in document and meant to be re-used.
These objects would only expose `fn write_ref` function, and document
would be responsible to render the actual contents of such objects.

### Checklist
<!-- 
Below is a fixed checklist to follow before requesting a code review. 
Please ensure that each item is completed to maintain code quality, 
improve maintainability, and streamline the review process. 
-->
- [x] Tests are added/updated.
- [x] Documentation is added/updated.
- [x] Self-review of code changes completed.

### Related Issues
None