feat: implement `Text` PDF object (#56)
### What?
This PR implements the `Text` PDF object.

### Why?
To enable text rendering on PDF pages, without which there's no great
use of this PDF rendering crate.

### How?
By introducing a `text` module in which we have 3 main components:
- TextTransform (Representing the positioning and size of the rendered
text)
- Text (Representing the Text object itself, encapsulating all the
required PDF data to make it work)
- TextBuilder (A builder, to conveniently create Text objects)

### Notes
- A slight adaptation of the `Font` object was required, adding a `Name`
name in it to represent the `Resources` reference.
- `PdfString` adaptation was required, switching the `PdfString` from a
stream based approach to a simplified literal based approach.

### Checklist
- [x] Tests are added/updated.
- [x] Documentation is added/updated.
- [x] Self-review of code changes completed.

### Related Issues
Closes #39