refactor: add type-safety to `ObjId` (#67)
### What?
Add type safety to `ObjId` so that it can't be misused.

### Why?
We want to enforce users to pass references to correct objects. Since we
control creation of `ObjId` instances, we can enforce that any given
instance points to the expected object (e.g. reference to a Font vs.
reference to an Image).

### How?
`PhantomData` is used as a marker for generic type on `ObjId`. Default
is `()` which would represent a reference to *any* PDF object. This
makes it possible to store references to different types of objects.
Also a `ObjId::cast()` method is implemented that easily converts
between references to any types.

### Checklist
- [x] Tests are added/updated.
- [x] Documentation is added/updated.
- [x] Self-review of code changes completed.