refactor: avoid allocating when writing formatted strings (#68)
### What?
We are writing strings using the `format!().as_bytes()` pattern. This
allocates a `String`, but it does not have to do it. This PR implements
small helper macro that avoids this allocation.

### Why?
We want to avoid allocation where it's not necessary.

### How?
Small adapter struct is implemented `WriteCounter` that implements
`std::fmt::Write` trait. This is then used with `fmt::write` function to
write contents for `format_args!()` into it and simply count number of
bytes written. This mechanism is then wrapped with small helper macro
that is similar to `std::write!()` and `format!()` macros, but avoids
allocating a `String`.

### Checklist
- [x] Tests are added/updated.
- [x] Documentation is added/updated.
- [x] Self-review of code changes completed.

### Related Issues
Closes #57