refactor: rename `Name` to `Identifier` (#69)
### What?
Renames the `Name` to `Identifier` (and all references to `Name`).

### Why?
Some keys in dictionary can literally be `Name`, so you could have a
`Name` with a value `Name`. This is very confusing. Instead, we're going
with the more standard name (pun not intended) for these kinds of values
- Identifier.

### Notes
`pdfgen_macros::const_names` is also renamed to
`pdfgen_macros::const_identifiers`.

### Checklist
- [x] Tests are added/updated.
- [x] Documentation is added/updated.
- [x] Self-review of code changes completed.

### Related Issues
Closes #64