feat: implement parsing of `Identifier` (#70)
### What?
This PR adds implementation for parsing of `Identifier`s (called `Name`
in PDF) to `pdfgen` crate.

### Why?
`Identifier` must be exposed as public type to users. But a single
invalid `Identifier` would make the whole generated PDF invalid. So we
need a mechanism to enable users to pass custom `Identifier`
(potentially generated dynamically) but still enforce correctness.

### How?
`impl FromStr for Identifier<String>` is added. This makes it possible
to create identifier by calling the
[`str::parse`](https://doc.rust-lang.org/std/primitive.str.html#method.parse)
method. Error is returned if the provided name is invalid.

### Notes
* Some other changes are also part of this PR, such as removing the
`Identifier::to_bytes` method and replacing it with non-allocating write
instead.
* This PR depends on #69.

### Checklist
- [x] Tests are added/updated.
- [x] Documentation is added/updated.
- [x] Self-review of code changes completed.

### Related Issues
Resolves #66