feat: implement color management (#71)
### What?
Implement initial and basic support for color management. This enables
choosing color for text for example.

### Why?
This is one of the required features, and enables stuff like text
styling.

### How?
An enum is implemented with currently only one variant for RGB colors.
This enum can then be used wherever we need coloring information.

### Notes
There are two concerns here: 

1. I chose enum approach, but it's not clear that this is the best
approach. It's convenient this way because we can see right away all
supported color spaces. However, we might have some methods on the enum
that aren't useful or supported in all variants. We might want to go
with `trait ColorSpace` instead (with some default methods implemented)
and define one struct per color space instead, and implement the
`ColorSpace` trait for each of them.
2. Currently `Color` enum is just slapped onto the `Text` object. This
should be part of styling information, but that can be refactored at a
later point.

### Checklist
- [x] Tests are added/updated.
- [x] Documentation is added/updated.
- [x] Self-review of code changes completed.

### Related Issues
Resolves #40